Move everything code-related into src
