+++ src/components/AppBar.tsx
@@ -1,4 +1,6 @@
 import { css } from '@emotion/react';
+import { useContext } from 'react';
+import { Context, UserContext } from '../store';
 import { Breakpoints } from '../styles/breakpoints';
 
 const AppBarStyles = {
@@ -14,6 +16,8 @@ const AppBarStyles = {
 };
 
 export const AppBar: React.FC = () => {
+  const { user, setUser } = useContext(UserContext);
+
   return (
     <aside css={AppBarStyles.Wrapper}>
       <div>AppBar</div>
