+++ src/styles/globals.ts
