+++ tests/app/App.spec.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from '../../pages/index';
+import App from '../../src/pages/index';
 
 describe('App', () => {
   it('renders without crashing', () => {
